locator: move small tablet POD/enum types into tablets_fwd.hh
tablet_id, global_tablet_id, tablet_task_type,
tablet_repair_incremental_mode and tablet_version_block (+
compare_tablet_version_block) have no dependency on the heavy
machinery in locator/tablets.hh (1241 lines, pulls in
seastar/core/reactor.hh, raft/raft.hh, locator/topology.hh) - they're
plain structs/enums, same family as tablet_replica/tablet_replica_set
which already live in tablets_fwd.hh.

Move them into tablets_fwd.hh alongside the existing fwd-declared
types, and repoint 9 includers that only ever needed this lightweight
subset (confirmed by grepping for tablet_map/tablet_metadata/
tablet_info/tablet_task_info usage - none found) at tablets_fwd.hh
instead of the full header: service/topology_coordinator.hh,
service/strong_consistency/{coordinator,groups_manager,state_machine}.hh,
service/topology_state_machine.cc, tasks/virtual_task_hint.hh,
tools/load_system_tablets.hh, cql3/query_options.hh,
db/system_keyspace.cc.

Targets full-build time via reduced transitive fan-out from
locator/tablets.hh (267 direct includers).
